Fix build errors on {i686,x86_64}-apple-darwin.

+++ test/test_stat.rs
@@ -38,11 +38,11 @@ fn assert_lstat_results(stat_result: Result<FileStat>) {
             assert!(stats.st_ino > 0);      // inode is positive integer, exact number machine dependent
             assert!(stats.st_mode > 0);     // must be positive integer
 
-            // st_mode is c_uint (u32) on Android and mode_t (u32) on all other
-            // *nix OSs; S_IFMT is mode_t (u16 on Android, u32 on all other
-            // *nix OSs). Therefore casting mode_t to u32 should be safe..
-            assert!(stats.st_mode & (posix88::S_IFMT as u32)
-                    == posix88::S_IFLNK as u32); // should be a link
+            // st_mode is c_uint (u32 on Android) while S_IFMT is mode_t
+            // (u16 on Android), and that will be a compile error.
+            // On other platforms they are the same (either both are u16 or u32).
+            assert!((stats.st_mode as usize) & (posix88::S_IFMT as usize)
+                    == posix88::S_IFLNK as usize); // should be a link
             assert!(stats.st_nlink == 1);   // there links created, must be 1
             // uid could be 0 for the `root` user. This quite possible when
             // the tests are being run on a rooted Android device.
